awg/requirements.txt
python-decouple==3.8
cryptography==44.0.1
